+++ include/reMath.h
@@ -1,11 +1,11 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
-// Name:		Razor Edge Math Library
-// File:		reMath.h
-// Project:		Razor Edge Classes
-// Description:	Cummulative include of Razor Edge math classes
-// Copyright:	Copyright © 2004++ REGLabs
-// Author:		Pavel Chikul
+// Name:        Razor Edge Math Library
+// File:        reMath.h
+// Project:     Razor Edge Classes
+// Description: Cummulative include of Razor Edge math classes
+// Copyright:   Copyright © 2004++ REGLabs
+// Author:      Pavel Chikul
 //
 //////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -14,7 +14,8 @@
 
 #include "reVec2d.h"
 #include "reVec3d.h"
-#include "reMatrix.h"
+#include "reMatrix3.h"
+#include "reMatrix4.h"
 #include "reQuaternion.h"
 #include "reMathUtil.h"