+++ include/reVec2d.h
@@ -1,10 +1,10 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
-// File:		reVec2d.h
-// Project:		Razor Edge Classes
-// Description:	Definition of Vec2d class (Math Module)
-// Copyright:	Copyright © 2004++ REGLabs
-// Author:		Pavel Chikul
+// File:        reVec2d.h
+// Project:     Razor Edge Classes
+// Description: Definition of Vec2d class (Math Module)
+// Copyright:   Copyright © 2004++ REGLabs
+// Author:      Pavel Chikul
 //
 //////////////////////////////////////////////////////////////////////////////////////////////////
 
@@ -13,113 +13,125 @@
 
 namespace re
 {
-	// Vec2d Class
+	// Vec2d Class.
 	class Vec2d
 	{
 	public:
-		// Constructors
+		// Constructors.
 		Vec2d();
 		Vec2d(const Vec2d& vector);
-		Vec2d(const Vec2d* vector);
-		Vec2d(float value);
+		explicit Vec2d(const Vec2d* vector);
+		explicit Vec2d(float value);
 		Vec2d(float xValue, float yValue);
-		Vec2d(const float* vector);
+		explicit Vec2d(const float* vector);
+		Vec2d(Vec2d&& vector) = default;
 
-		// Destructor
+		// Destructor.
 		virtual ~Vec2d() = default;
 
 	public:
-		// Set vector data
+		// Set vector data.
 		void set(const Vec2d& vector);
 		void set(float value);
 		void set(float xValue, float yValue);
 		void set(const float* vector);
 
-		// Get vector magnitutde
+		// Get vector magnitutde.
 		float length() const;
 
-		// Parallel vectors check
+		// Parallel vectors check.
 		bool isParallel(const Vec2d& vector) const;
 
-		// Normalize vector
+		// Normalize vector.
 		void normalize();
 
-		// Nagate vector
+		// Nagate vector.
 		void negate();
 
-		// Calculate cross product and return the result
+		// Calculate cross product and return the result.
 		// Note: The cross-product of 2 vectors is only defined in 3D and 7D spaces.
 		// This method computes the z-component of 2 vectors lying on the xy-plane.
 		float cross(const Vec2d& vector) const;
 
-		// Calculate dot product and return the result
+		// Calculate dot product and return the result.
 		float dot(const Vec2d& vector) const;
 
-		// Comparison operators
-		//---------------------
+		// Comparison operators.
+		//----------------------
 
-		// Equal to operator - performs by value comparison of vector data
+		// Equal to operator - performs by value comparison of vector data.
 		bool operator == (const Vec2d& vector) const;
 
-		// Not equal to operator - performs by value comparison of vector data
+		// Not equal to operator - performs by value comparison of vector data.
 		bool operator != (const Vec2d& vector) const;
 
-		// Assignment operators
-		//---------------------
+		// Assignment operators.
+		//----------------------
 
-		// Copy assignment operator
+		// Copy assignment operator.
 		Vec2d& operator = (const Vec2d& vector);
 
+		// Move assignment operator.
+		Vec2d& operator = (Vec2d&& vector) = default;
 
-		// Arithmetic operators
-		//---------------------
+		// Arithmetic operators.
+		//----------------------
 
-		// Unary minus - Negates vector
+		// Unary minus - Negates vector.
 		Vec2d operator - () const;
 
-		// Returns result of two vector addition
+		// Returns result of two vector addition.
 		Vec2d operator + (const Vec2d& vector) const;
 
-		// Returns result of two vector substraction
+		// Returns result of two vector substraction.
 		Vec2d operator - (const Vec2d& vector) const;
 
-		// Multiplies vector elements by value
+		// Multiplies vector elements by value.
 		friend Vec2d operator * (float value, const Vec2d& vector);
 
-		// Multiplies vector elements by value
+		// Multiplies vector elements by value.
 		friend Vec2d operator * (const Vec2d& vector, float value);
 
 
-		// Compound assignment operators
-		//------------------------------
+		// Compound assignment operators.
+		//-------------------------------
 
-		// Vector addition
+		// Vector addition.
 		void operator += (const Vec2d& vector);
 
-		// Adds value to every vector component
+		// Adds value to every vector component.
 		void operator += (float value);
 
-		// Vector substraction
+		// Vector substraction.
 		void operator -= (const Vec2d& vector);
 
-		// Substracts value from every vector component
+		// Substracts value from every vector component.
 		void operator -= (float value);
 
-		// Multiplies every vector component by value
+		// Multiplies every vector component by value.
 		void operator *= (float value);
 
-		// Divides every vector component by value
+		// Divides every vector component by value.
 		void operator /= (float value);
 
 
-		// Conversion operators
+		// Conversion operators.
+		//----------------------
+
+		// Returns a pointer to vector data.
+		explicit operator float* ();
+
+		// Returns a constant pointer to vector data.
+		explicit operator const float* () const;
+
+		// Subscript operators.
 		//---------------------
 
-		// Returns a pointer to vector data
-		operator float* ();
+		// Data array access operator.
+		float& operator [] (size_t index);
 
-		// Returns a constant pointer to vector data
-		operator const float* () const;
+		// Constant data array access operator.
+		const float& operator [] (size_t index) const;
 
 	public:
 		union