+++ source/reMathUtil.cpp
@@ -1,16 +1,15 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
-// File:		reMathUtil.cpp
-// Project:		Razor Edge Classes
-// Description:	Implementation of Razor Edge math utility functions (Math Module)
-// Copyright:	Copyright © 2004++ REGLabs
-// Author:		Pavel Chikul
+// File:        reMathUtil.cpp
+// Project:     Razor Edge Classes
+// Description: Implementation of Razor Edge math utility functions (Math Module)
+// Copyright:   Copyright © 2004++ REGLabs
+// Author:      Pavel Chikul
 //
 //////////////////////////////////////////////////////////////////////////////////////////////////
 
 #include "reMathUtil.h"
 #include "reMath.h"
-#include <string.h>
 #include <math.h>
 
 float re::toRadians(float degrees)
@@ -57,18 +56,56 @@ unsigned char re::getLowNibble(unsigned char byte)
 
 
 
-re::Matrix re::perspective(float fovy, float aspect, float zNear, float zFar)
+re::Matrix4 re::perspective(float fovy, float aspect, float zNear, float zFar)
 {
-	float matrix[16];
-	memset(matrix, 0, sizeof(matrix));
+	const float tanHalfFovy = 1.f / tanf(toRadians(fovy) / 2.f);
 
-	float cotngentHalfFovy = float(1.f / (tan(toRadians(fovy) / 2.f)));
-
-	matrix[0] = cotngentHalfFovy / aspect;
-	matrix[5] = cotngentHalfFovy;
+	Matrix4 matrix;
+	matrix[0] = tanHalfFovy / aspect;
+	matrix[5] = tanHalfFovy;
 	matrix[10] = (zFar + zNear) / (zNear - zFar);
 	matrix[11] = -1.f;
 	matrix[14] = (2.f * zFar * zNear) / (zNear - zFar);
+	return matrix;
+}
 
-	return Matrix(matrix);
+re::Matrix4 re::lookAt(const Vec3d& eye, const Vec3d& center, const Vec3d& up)
+{
+	// Calculate forward vector
+	Vec3d forward(center - eye);
+	forward.normalize();
+
+	// Normalize up vector
+	Vec3d upVector(up);
+	upVector.normalize();
+
+	// Get the perpendicular "side" vector
+	Vec3d side(forward.cross(up));
+	side.normalize();
+
+	// Recompute up as "side x forward"
+	upVector = side.cross(forward);
+
+	Matrix4 matrix2, result;
+	
+	// Column #1
+	matrix2[0] = side.x;
+	matrix2[4] = side.y;
+	matrix2[8] = side.z;
+
+	// Column #2
+	matrix2[1] = upVector.x;
+	matrix2[5] = upVector.y;
+	matrix2[9] = upVector.z;
+
+	// Column #3
+	matrix2[2] = -forward.x;
+	matrix2[6] = -forward.y;
+	matrix2[10] = -forward.z;
+
+	// Set eye translation
+	result *= matrix2;
+	result.setTranslation(-eye);
+	
+	return result;
 }